+++ Source/PluginProcessor.cpp
@@ -5,7 +5,7 @@
 
   ==============================================================================
 */
-
+#include "FaustEchoEffect.h"
 #include "PluginProcessor.h"
 #include "PluginEditor.h"
 
@@ -13,8 +13,8 @@
 juce::AudioProcessorValueTreeState::ParameterLayout createParameterLayout()
 {
     juce::AudioProcessorValueTreeState::ParameterLayout params;
-    std::string fxNamesStr[4]={"Chopper","Filter","Echo","Crusher"};
-    juce::StringArray fxNames=juce::StringArray({"Chopper","Filter","Echo","Crusher"});
+    std::string fxNamesStr[4]={"Chopper","Echo","Filter","Crusher"};
+    juce::StringArray fxNames=juce::StringArray({"Chopper","Echo","Filter","Crusher"});
  
     params.add (std::make_unique<juce::AudioParameterFloat> ("drywet","Dry/Wet", 0.0f, 1.0f, 1.0f));
     params.add (std::make_unique<juce::AudioParameterFloat> ("outgain","Out Gain",0.0f, 2.0f, 1.0f));
@@ -48,6 +48,16 @@ FxseqAudioProcessor::FxseqAudioProcessor()
        pluginParameters (*this, nullptr, "PARAMETERS", createParameterLayout())
 {
     initAllPatterns();
+    
+    filter.reset();
+    //filter.setMode(juce::dsp::LadderFilter<float>::Mode::LowPass);
+    filter.setCutoffFrequencyHz(20000.0f);
+    filter.setResonance(0.0f);
+
+    fx1dryWetMixer.setWetMixProportion(1.0f);fx1dryWetMixer.setWetLatency(0.0f);
+    fx2dryWetMixer.setWetMixProportion(1.0f);fx2dryWetMixer.setWetLatency(0.0f);
+    fx3dryWetMixer.setWetMixProportion(1.0f);fx3dryWetMixer.setWetLatency(0.0f);
+    fx4dryWetMixer.setWetMixProportion(1.0f);fx4dryWetMixer.setWetLatency(0.0f);
 }
 
 FxseqAudioProcessor::~FxseqAudioProcessor()
@@ -118,15 +128,41 @@ void FxseqAudioProcessor::changeProgramName (int index, const juce::String& newN
 
 //==============================================================================
 void FxseqAudioProcessor::prepareToPlay (double sampleRate, int samplesPerBlock)
-{
-    // Use this method as the place to do any pre-playback
-    // initialisation that you need..
+{   
+    // FAUST ECHO
+    echoDSP = new mydsp();
+    echoDSP->init(sampleRate);
+    echoUI = new MapUI();
+    echoDSP->buildUserInterface(echoUI);
+    echoInputs = new float*[2];
+    echoOutputs = new float*[2];
+    for (int channel = 0; channel < 2; ++channel) {
+        echoInputs[channel] = new float[samplesPerBlock];
+        echoOutputs[channel] = new float[samplesPerBlock];
+    }
+
+    filter.reset();
+
+    fx1drybuffer.setSize(2, samplesPerBlock);
+    fx2drybuffer.setSize(2, samplesPerBlock);
+    fx3drybuffer.setSize(2, samplesPerBlock);
+    fx4dryBuffer.setSize(2, samplesPerBlock); 
+
+    //filter.setSampleRate(sampleRate);
+    //filter.prepare({ sampleRate, (double)maximumBlockSize, (double)getTotalNumInputChannels() });
 }
 
 void FxseqAudioProcessor::releaseResources()
 {
-    // When playback stops, you can use this as an opportunity to free up any
-    // spare memory, etc.
+    // FAUST ECHO
+    delete echoDSP;
+    delete echoUI;
+    for (int channel = 0; channel < 2; ++channel) {
+        delete[] echoInputs[channel];
+        delete[] echoOutputs[channel];
+    }
+    delete [] echoInputs;
+    delete [] echoOutputs;
 }
 
 #ifndef JucePlugin_PreferredChannelConfigurations
@@ -169,36 +205,61 @@ void FxseqAudioProcessor::processBlock (juce::AudioBuffer<float>& buffer, juce::
         
         for (int i=0;i<sequencerCount;i++) // FOR EACH SEQUENCER
         {   
-            sequencerClockMult[i]=(int)pow(2,(pluginParameters.getParameter("seq"+std::__cxx11::to_string(i+1)+"_clockDiv")->getValue()*7));     
+            sequencerClockMult[i]=(int)pow(2,(pluginParameters.getParameter("seq"+std::__cxx11::to_string(i+1)+"_clockDiv")->getValue()*7));    // get clock mult 
             sequencerPositions[i]=int(std::floor( std::modf(ppq/sequencerClockMult[i],&unused)*(16*resolution))); // compute position on pattern
             if (not (bool)pluginParameters.getParameter("sequencerMode")->getValue()){  // PATTERN MODE          
-                selected_pattern[i]=(int)getParameterValue("seq"+std::__cxx11::to_string(i+1)+"_pattern"); 
+                selected_pattern[i]=(int)getParameterValue("seq"+std::__cxx11::to_string(i+1)+"_pattern"); //get selected pattern
             } else {                                                                   // SEQUENCE MODE
                 int selectedSequence=(int)(pluginParameters.getParameter("sequenceNumber")->getValue()*15); // get selected sequence (SEQUENCE MODE ONLY)                 
                 int sequencePosition;
                 sequenceLength=(int)getParameterValue("sequenceLength")+2;
                 if (ppq >= (float)sequenceLength ) {sequencePosition=int(std::floor(ppq/greatestClockMult)-sequenceLength*std::floor(ppq/(sequenceLength*greatestClockMult)));} else {sequenceLength=std::floor(ppq/greatestClockMult);}         // compute positon on sequence (SEQUENCE MODE ONLY)             
                 
-                if ((sequencePosition >15) or (sequencePosition<0)) {sequencePosition=0;}
+                sequencePosition=std::clamp(sequencePosition,0,15); // safety 
                 selected_pattern[i]=sequences[selectedSequence][sequencePosition]; // update selected pattern
             }
             lastFxDepths[i]=fxDepths_smoothed[i].getNextValue(); // get smoothed value from previous value
-            fxDepths_smoothed[i].setTargetValue(gainPatterns[i][selected_pattern[i]][sequencerPositions[i]]);// get data from pattern and set it for smoothed value target );           
+            fxDepths_smoothed[i].setTargetValue(gainPatterns[i][selected_pattern[i]][sequencerPositions[i]]);// get data from pattern and set it for smoothed value target            
         }
         
        
     }
-   
+    /*
     for (auto i = totalNumInputChannels; i < totalNumOutputChannels; ++i)
+    {
         buffer.clear (i, 0, buffer.getNumSamples());
+    }
 
-    for (int channel = 0; channel < totalNumInputChannels; ++channel)
+    for (auto i = totalNumInputChannels; i < totalNumOutputChannels; ++i) {  // Faust echo 
+            for (int j = 0; j < buffer.getNumSamples(); j++) {echoInputs[i][j] = *buffer.getWritePointer(i,j); }
+    }
+    echoDSP->compute(buffer.getNumSamples(),echoInputs,echoOutputs); // Faust echo
+
+    for (int channel = 0; channel < totalNumInputChannels; ++channel) 
     {
         auto* channelData = buffer.getWritePointer (channel);
-        //buffer.applyGain(fxDepths_smoothed[0].getNextValue());
-        buffer.applyGainRamp(channel,0,buffer.getNumSamples(),lastFxDepths[0],fxDepths_smoothed[0].getNextValue());
-        // ..do something to the data...
-    }
+        //buffer.applyGain(fxDepths_smoothed[0].getNextValue()); // Chopper gain (clicks)
+        buffer.applyGainRamp(channel,0,buffer.getNumSamples(),lastFxDepths[0],fxDepths_smoothed[0].getNextValue());// Chopper gain (ramp)
+        //for (int i = 0; i < buffer.getNumSamples(); i++){ *buffer.getWritePointer(channel,i) = echoOutputs[channel][i];       } // Faust echo        
+
+    }*/
+    
+    float fx3mix=getParameterValue("Filter_dry/wet");
+          
+    chopper_process(buffer,fx1drybuffer);//FX1  
+  
+    echo_process(buffer); // FX2
+
+    for (int channel = 0; channel < totalNumOutputChannels; ++channel) { //FX3
+        juce::dsp::AudioBlock<float> audioBlock(buffer);   // Ladder filter
+        juce::dsp::ProcessContextReplacing<float> context(audioBlock);// Ladder filter
+        //filter.process(context);// Ladder filter
+    }   
+
+    bitcrush_process(buffer); // FX4
+    
+
+  
 }
 
 //==============================================================================
@@ -282,4 +343,95 @@ float FxseqAudioProcessor::getParameterValue(std::string paramName)
 {
     auto rg=pluginParameters.getParameter(paramName)->getNormalisableRange().end-pluginParameters.getParameter(paramName)->getNormalisableRange().start;
     return pluginParameters.getParameter(paramName)->getValue()*rg;
+}
+
+void FxseqAudioProcessor::updateEffectProgramParameter(int effectIndex,int programIndex,int parameterIndex,float parameterValue)
+{
+    fxPrograms[effectIndex][programIndex][parameterIndex]=parameterValue;
+}
+//////////////////////////// FX CALLBACKS ////////////////////////////////////////////////////////////////////////////////////////////////////////////
+void FxseqAudioProcessor::chopper_process(juce::AudioBuffer<float>& buffer,juce::AudioBuffer<float>& dryBuffer)
+{
+    float fx1mix=getParameterValue("Chopper_dry/wet");
+
+    for (int channel = 0; channel < 2; ++channel) {
+        dryBuffer.copyFrom(channel, 0, buffer, channel, 0, buffer.getNumSamples());// copy dry input 
+        buffer.applyGainRamp(channel,0,buffer.getNumSamples(),lastFxDepths[0],fxDepths_smoothed[0].getNextValue());  // chopper         
+    }
+    float wet = fx1mix;
+    float dry = 1.0f - fx1mix;
+
+    juce::FloatVectorOperations::multiply(dryBuffer.getWritePointer(0), dry, buffer.getNumSamples());
+	juce::FloatVectorOperations::multiply(dryBuffer.getWritePointer(1), dry, buffer.getNumSamples());
+
+	juce::FloatVectorOperations::multiply(buffer.getWritePointer(0), wet, buffer.getNumSamples());
+	juce::FloatVectorOperations::multiply(buffer.getWritePointer(1), wet, buffer.getNumSamples());
+
+	juce::FloatVectorOperations::add(buffer.getWritePointer(0), dryBuffer.getReadPointer(0), buffer.getNumSamples());
+	juce::FloatVectorOperations::add(buffer.getWritePointer(1), dryBuffer.getReadPointer(1), buffer.getNumSamples());    
+}
+
+void FxseqAudioProcessor::echo_process(juce::AudioBuffer<float>& buffer)
+{
+    float fx2mix=getParameterValue("Echo_dry/wet");
+    float echoMix=fxDepths_smoothed[1].getNextValue();
+
+    for (int channel = 0; channel < 2; ++channel) {
+        for (int i = 0; i < buffer.getNumSamples(); i++) { 
+                echoInputs[channel][i] = *buffer.getWritePointer(channel,i); //Echo
+            }
+    }
+    echoDSP->compute(buffer.getNumSamples(),echoInputs,echoOutputs);
+    for (int channel = 0; channel < 2; ++channel) {
+        for (int i = 0; i < buffer.getNumSamples(); i++){
+            *buffer.getWritePointer(channel,i) =echoOutputs[channel][i]*echoMix*fx2mix+(1-echoMix*fx2mix) * echoInputs[channel][i]; // echo  
+        }
+    }
+}
+
+void FxseqAudioProcessor::bitcrush_process(juce::AudioBuffer<float>& buffer)
+{
+    int reduction = int(fxPrograms[1][patterns[3][selected_pattern[3]][int(sequencerPositions[3]/resolution)]][0]);// FX4
+
+    float fx4mix=getParameterValue("Crusher_dry/wet");    
+    float crusherMix=std::clamp(fxDepths_smoothed[3].getNextValue(),0.0f,1.0f);//clamp(fxDepths_smoothed[3].getNextValue(),0,1);
+    for (int channel = 0; channel < 2; ++channel) {
+        float* channelData = buffer.getWritePointer(channel);
+        for (int i = 0; i < buffer.getNumSamples(); i++){
+ 
+             if (reduction > 1) {if (i%reduction != 0) channelData[i] = crusherMix*fx4mix*channelData[i - i%reduction]+(1-crusherMix*fx4mix)*channelData[i];} // sample rate reduction
+        }
+    }
+}
+
+//////////////////////////// FX /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+void FxseqAudioProcessor::echo_setDelay(float delay)
+{
+    echoUI->setParamValue("delay",delay);
+}
+
+void FxseqAudioProcessor::echo_setFeedback(float feedback)
+{
+    echoUI->setParamValue("feedback",feedback);
+}
+
+float FxseqAudioProcessor::bitcrusher_bitReduction(float sample, float bitReduction)
+{
+    float maxValue = pow(2, bitReduction) - 1;
+    float crushedSample = round(sample * maxValue) / maxValue;
+    return crushedSample;
+}
+
+float FxseqAudioProcessor::bitcrusher_sampleRateReduction(float sample, int reductionFactor)
+{
+    if (bitcrusher_count % reductionFactor == 0)
+    {
+        bitcrusher_count = 0;
+        return sample;
+    }
+    else
+    {
+        bitcrusher_count++;
+        return 0.0f;
+    }
 }